+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-### Version 4.29.0 (xxth June 2021)
+### Version 4.29.0 (11th June 2021)
 #### Added
 - Added support for null safety (thanks to @blaueeiner).
-- Added data residency feature. You can choose this setting by setting `urlStrategy` member of `AdjustConfig` instance to `AdjustConfig.DataResidencyEU` (for EU data residency region), `AdjustConfig.DataResidencyTR` (for TR data residency region) or `AdjustConfig.DataResidencyUS` value (for US data residency region).
+- [beta] Added data residency feature. You can choose this setting by setting `urlStrategy` member of `AdjustConfig` instance to `AdjustConfig.DataResidencyEU` (for EU data residency region), `AdjustConfig.DataResidencyTR` (for TR data residency region) or `AdjustConfig.DataResidencyUS` value (for US data residency region).
+- Added `trackAdRevenueNew` method to `Adjust` interface to allow tracking of ad revenue by passing `AdjustAdRevenue` object as parameter.
+- Added support for `AppLovin MAX` ad revenue tracking.
+- Added `conversionValueUpdatedCallback` member to `AdjustConfig` which can be used to set a callback which will get information when Adjust SDK updates conversion value for the user.
 - Added preinstall tracking with usage of system installer receiver on Android platform.
 
 #### Fixed